Using PERL_LOCAL_LIB_ROOT if set.
This makes the module installable when using plenv / local::lib alone,
which then solves the issue of binaries and module code erroneously
being put into the perl installation directory. The perl installation
directory might not even be writable by the installing user, so this is
needed if this module is to be used in a shared environment.

See also:

    https://rt.cpan.org/Public/Bug/Display.html?id=116956
    https://github.com/jeteve/Alien-ImageMagick/issues/4